Move DrawerFolder and ReminderContents to project/taskDetails
